Add 0-4.88fm embedding option

+++ submit/JS_pp200_signal/pass4jet_embed/condor/restart_held.sh
@@ -6,6 +6,9 @@ if [ $# -eq 0 ]
 fi
 echo $1
 variable=$1
+run=7
+fm=0_488fm
+
 #exit 0
 condor_q | grep ' H ' | grep run_pass4jet_embed_js.sh | grep ${variable} > bla
 
@@ -13,16 +16,10 @@ condor_q | grep ' H ' | grep run_pass4jet_embed_js.sh | grep ${variable} > bla
 
 for i in `cat bla| awk '{print $1}'`; do condor_rm $i; done
 
-if [ -f tmplist ]
-then
-rm tmplist
-fi
-for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. '{print "-0000000007-"$1".job"}'`; do echo $i >> tmplist ; done
+[ -f tmplist ] && rm tmplist
+for i in `cat bla | awk '{print $12}' | awk -F- '{print $3}' | awk -F. -v run=${run} '{print "-000000000"run"-"$1".job"}'`; do echo $i >> tmplist ; done
 
-if [ -f sedlist ]
-then
-rm sedlist
-fi
-for i in `cat tmplist`; do echo log/${variable}/condor_${variable}$i >> sedlist; done
+[ -f sedlist ] && rm sedlist
+for i in `cat tmplist`; do echo log/${fm}/run${run}/${variable}/condor_${variable}$i >> sedlist; done
 for i in `cat sedlist`; do  sed -i 's/6144MB/12288MB/' $i; echo $i; done
 for i in `cat sedlist`; do condor_submit $i; done
